Rename `Map` component to `DynamicMap`, to better signal its complexity
And add a describing comment
