+++ src/components/DynamicMap/NightShadow.tsx
