+++ src/pages/MapPage.tsx
@@ -1,8 +1,8 @@
 import { ErrorBoundary } from 'react-error-boundary';
 import { ModalsProvider } from '@mantine/modals';
 
+import { DynamicMap } from '@/components/DynamicMap/DynamicMap';
 import { fallbackRender } from '@/components/ErrorFallback/FallbackRender';
-import { Map } from '@/components/Map/Map';
 import { ConnectionErrorOverlay } from '@/windowmanagement/ConnectionErrorOverlay';
 import { Window } from '@/windowmanagement/Window/Window';
 
@@ -15,7 +15,7 @@ export function MapPage() {
       <ModalsProvider>
         <ConnectionErrorOverlay />
         <Window h={'100vh'} w={'100vw'}>
-          <Map iconSize={55} coneHeight={55} coneWidth={70} />
+          <DynamicMap iconSize={55} coneHeight={55} coneWidth={70} />
         </Window>
       </ModalsProvider>
     </ErrorBoundary>
